Reworks locales for clarity on rich vs text, errors, etc

+++ prism-bukkit/src/main/java/network/darkhelmet/prism/commands/PurgeCommand.java
@@ -29,13 +29,14 @@ import dev.triumphteam.cmd.core.annotation.NamedArguments;
 import dev.triumphteam.cmd.core.annotation.SubCommand;
 import dev.triumphteam.cmd.core.argument.named.Arguments;
 
+import java.util.Optional;
+
 import network.darkhelmet.prism.api.activities.ActivityQuery;
 import network.darkhelmet.prism.api.services.purges.IPurgeQueue;
 import network.darkhelmet.prism.loader.services.configuration.ConfigurationService;
 import network.darkhelmet.prism.services.messages.MessageService;
 import network.darkhelmet.prism.services.purge.PurgeService;
 import network.darkhelmet.prism.services.query.QueryService;
-import network.darkhelmet.prism.services.translation.TranslationKey;
 
 import org.bukkit.command.CommandSender;
 
@@ -92,21 +93,23 @@ public class PurgeCommand extends BaseCommand {
     @Permission("prism.admin")
     public void onPurge(final CommandSender sender, final Arguments arguments) {
         if (!purgeService.queueFree()) {
-            messageService.error(sender, new TranslationKey("purge-queue-not-free"));
+            messageService.errorPurgeQueryNotFree(sender);
         }
 
-        final ActivityQuery query = queryService
-            .queryFromArguments(sender, arguments)
-            .limit(configurationService.prismConfig().purges().limit())
-            .build();
+        Optional<ActivityQuery.ActivityQueryBuilder> builder = queryService.queryFromArguments(sender, arguments);
+        if (builder.isPresent()) {
+            final ActivityQuery query = builder.get()
+                .limit(configurationService.prismConfig().purges().limit())
+                .build();
 
-        IPurgeQueue purgeQueue = purgeService.newQueue(result -> {
-            messageService.purgeCycle(sender, result);
-        }, result -> messageService.purgeComplete(sender, result.deleted()));
+            IPurgeQueue purgeQueue = purgeService.newQueue(result -> {
+                messageService.purgeCycle(sender, result);
+            }, result -> messageService.purgeComplete(sender, result.deleted()));
 
-        purgeQueue.add(query);
-        purgeQueue.start();
+            purgeQueue.add(query);
+            purgeQueue.start();
 
-        messageService.purgeStarting(sender);
+            messageService.purgeStarting(sender);
+        }
     }
 }
